refactor: database logic into mixins and enhance documentation
- Split TimelinkDatabase into specialized mixins (Kleio, Metadata, Query, Views)
- Add comprehensive Google-style docstrings to all database-related modules
- Fix linting errors and unused imports in the API package
- Update test reference files for consistency
