+++ tests/timelink-home/projects/test-project/sources/reference_sources/devassas/dev1692.cli
@@ -1,4 +1,4 @@
-kleio$gacto2.str/translations=245
+kleio$gacto2.str/translations=246
 
    property$multiple-entry-flag/124#ascii for pipe char
    fonte$dev1692/1692/devassa/loc=auc III-d,5,2,135
