+++ tests/timelink-home/projects/test-project/sources/reference_sources/issues/issue48/issue48.cli
@@ -1,4 +1,4 @@
-kleio$/translations=102
+kleio$/translations=103
    fonte$issue48/1982/bibliografia
 
       carta$deh-ca-1645-b-2-a/0/0/1643/?/refere=1645/resumo=LAVP da China do Norte dos annos 1643, 1644, 1645/obs="BNL 722, 298-328 715-755; J 49-V-13, 209-211; Cy nan, ACAL 2lbis fasc. 21, 761"
