+++ tests/timelink-home/projects/test-project/sources/reference_sources/issues/issue48/issue48b.cli
@@ -1,4 +1,4 @@
-kleio$gacto2str/translations=166
+kleio$gacto2str/translations=167
 
    property$multiple-entry-flag/124#ascii for pipe char/id=root-pro1-78
    link$uc-china/"https://timelink.uc.pt/mhk/china/id/$1"/id=root-lin1-78
