+++ tests/timelink-home/projects/test-project/sources/reference_sources/issues/multiple_databases/dev1692.cli
@@ -1,4 +1,4 @@
-kleio$gacto2.str/translations=10
+kleio$gacto2.str/translations=11
 
    property$multiple-entry-flag/124#ascii for pipe char
    fonte$dev1692/1692/devassa/loc=auc III-d,5,2,135/obs="short version for testing"
