+++ tests/timelink-home/projects/test-project/sources/reference_sources/linked_data/dehergne-a-partial.cli
@@ -1,4 +1,4 @@
-kleio$gacto2.str/translations=821
+kleio$gacto2.str/translations=822
    link$wikidata/"https://www.wikidata.org/wiki/$1"/id=root-lin1-231
    fonte$dehergne-a-partial/1973/Dicionário Biográfico/loc=Online archive.org:details:bhsi37/obs=Dehergne, Joseph, Répertoire des Jésuites de Chine, de 1542 à 1800, 1973. Letra A - Entradas relacionadas com Coimbra.
