+++ tests/timelink-home/projects/test-project/sources/reference_sources/linked_data/dehergne-locations-1644.cli
@@ -1,4 +1,4 @@
-kleio$gacto2.str/translations=715
+kleio$gacto2.str/translations=716
    link$wikidata/"https://www.wikidata.org/wiki/$1"/id=root-lin2-44
    link$geonames/"https://www.geonames.org/$1"/id=root-lin2-45
    fonte$dehergne-locations-1644/1644/geoinformation/substitui=dehergne-map-1644/obs="""
