+++ tests/timelink-home/projects/test-project/sources/reference_sources/rentities/sameas-tests.cli
@@ -1,4 +1,4 @@
-kleio$/translations=681
+kleio$/translations=685
    link$wikidata/"https://www.wikidata.org/wiki/$1"/id=root-lin1-4
    fonte$sameas-tests/obs="""
