+++ tests/timelink-home/projects/test-project/sources/reference_sources/test-attr-dates/test-attr-dates.cli
@@ -1,4 +1,4 @@
-kleio$gacto2.str/Generated from Archeevo export file/translations=65
+kleio$gacto2.str/Generated from Archeevo export file/translations=66
    fonte$test-attr-dates/"PT/AUC/ELU/UC-AUC/B/001-001"/data=1537-00-00:1919-11-14/loc=Arquivo da Universidade de Coimbra/tipo=teste
 
       lista$lista-test-atr-dates/11/02/2020/loc=A/obs=Exemplo de registos com datas nos atributos
