+++ tests/timelink-home/projects/test-project/sources/reference_sources/varia/auc-alunos.cli
@@ -1,6 +1,6 @@
 kleio$gacto2.str/"""
       Generated from Archeevo export file:  database/other_exports/ficheiro-alunos-auc/PT-AUC-ELU-UC-AUC-B-001-001.CSV dated 2022-02-03 14:43
-  """/translations=82
+  """/translations=83
    fonte$auc-alunos/"PT/AUC/ELU/UC-AUC/B/001-001"/data=1537-00-00:1919-11-14/loc=Arquivo da Universidade de Coimbra/substitui=auc-alunos-264605-A-140337-140771/tipo=UC-AUC-ficheiro-alunos/obs="""
        Âmbito e conteúdo
