+++ tests/timelink-home/projects/test-project/structures/reference_sources/issues/issue48/sources.str.yaml
@@ -1,4 +1,4 @@
-- path: /kleio-home/structures/reference_sources/issues/issue48/sources.str
+- path: /kleio-home/projects/test-project/structures/reference_sources/issues/issue48/sources.str
 - group:
     also:
     - structure
