+++ timelink/api/database_postgres.py
@@ -1,15 +1,37 @@
-"""This module provides functions to manage a PostgreSQL server running in a Docker container.
-It includes functions to check if Docker and PostgreSQL are running, retrieve PostgreSQL
-container details such as user and password, construct a PostgreSQL URL, and start a PostgreSQL
-server in a Docker container.
+"""PostgreSQL database management for Timelink.
+
+This module provides functions to manage a PostgreSQL server running in a Docker container.
+It includes utilities to check if Docker and PostgreSQL are running, retrieve PostgreSQL
+container details such as user credentials and connection parameters, construct PostgreSQL
+connection URLs, and start a PostgreSQL server in a Docker container with custom configuration.
 
 Functions:
-- is_postgres_running: Check if PostgreSQL is running in Docker.
-- get_postgres_container: Get the PostgreSQL container.
-- get_postgres_container_pwd: Get the PostgreSQL container password.
-- get_postgres_container_user: Get the PostgreSQL container user.
-- get_postgres_url: Get the PostgreSQL URL for a given database name.
-- start_postgres_server: Start a PostgreSQL server in Docker."""
+    is_postgres_running() -> bool:
+        Check if PostgreSQL is running in Docker.
+
+    get_postgres_container() -> docker.models.containers.Container:
+        Get the PostgreSQL Docker container object.
+
+    get_postgres_container_pwd() -> str:
+        Get the PostgreSQL container password from environment variables.
+
+    get_postgres_container_user() -> str:
+        Get the PostgreSQL container user from environment variables.
+
+    get_postgres_url(dbname: str) -> str:
+        Construct a PostgreSQL connection URL for a given database name.
+
+    start_postgres_server(dbname: str | None = "timelink", dbuser: str | None = "timelink",
+                         dbpass: str | None = None, image: str | None = "postgres",
+                         version: str | None = "latest") -> docker.models.containers.Container:
+        Start a PostgreSQL server in Docker with the specified configuration.
+
+    get_postgres_dbnames() -> list[str]:
+        Get the list of non-template databases from a running PostgreSQL server.
+
+    is_valid_postgres_db_name(db_name: str) -> bool:
+        Validate a PostgreSQL database name according to PostgreSQL naming rules.
+"""
 
 import re
 import time
@@ -29,7 +51,14 @@ postgres_container: docker.models.containers.Container = None
 
 
 def is_postgres_running():
-    """Check if postgres is running in docker"""
+    """Check if PostgreSQL is running in a Docker container.
+
+    Returns:
+        bool: True if at least one PostgreSQL container is running, False otherwise.
+
+    Note:
+        If Docker is not running, a warning is issued and False is returned.
+    """
 
     if not is_docker_running():
         warnings.warn("Docker is not running", stacklevel=2)
@@ -47,9 +76,14 @@ def is_postgres_running():
 
 
 def get_postgres_container() -> docker.models.containers.Container:
-    """Get the postgres container
+    """Get the PostgreSQL Docker container object.
+
     Returns:
-        docker.models.containers.Container: the postgres container
+        docker.models.containers.Container: The PostgreSQL container object.
+
+    Raises:
+        RuntimeError: If Docker is not running.
+        IndexError: If no PostgreSQL container is found.
     """
 
     if not is_docker_running():
@@ -65,9 +99,16 @@ def get_postgres_container() -> docker.models.containers.Container:
 
 
 def get_postgres_container_pwd() -> str:
-    """Get the postgres container password
+    """Get the PostgreSQL container password from environment variables.
+
     Returns:
-        str: the postgres container password
+        str: The PostgreSQL container password.
+
+    Raises:
+        RuntimeError: If Docker is not running.
+
+    Note:
+        Returns None if PostgreSQL is not running or password is not found.
     """
     if not is_docker_running():
         raise RuntimeError("Docker is not running")
@@ -83,9 +124,16 @@ def get_postgres_container_pwd() -> str:
 
 
 def get_postgres_container_user() -> str:
-    """Get the postgres container user
+    """Get the PostgreSQL container user from environment variables.
+
     Returns:
-        str: the postgres container user
+        str: The PostgreSQL container user.
+
+    Raises:
+        RuntimeError: If Docker is not running.
+
+    Note:
+        Returns None if PostgreSQL is not running or user is not found.
     """
     if not is_docker_running():
         raise RuntimeError("Docker is not running")
@@ -101,7 +149,15 @@ def get_postgres_container_user() -> str:
 
 
 def get_postgres_url(dbname: str) -> str:
-    """Get the postgres url for dbname"""
+    """Construct a PostgreSQL connection URL for a given database name.
+
+    Args:
+        dbname (str): The name of the database.
+
+    Returns:
+        str: A PostgreSQL connection URL in the format:
+            postgresql://user:password@localhost:5432/dbname
+    """
     usr = get_postgres_container_user()
     pwd = get_postgres_container_pwd()
     return f"postgresql://{usr}:{pwd}@localhost:5432/{dbname}"
@@ -114,12 +170,23 @@ def start_postgres_server(
     image: str | None = "postgres",
     version: str | None = "latest",
 ):
-    """Starts a postgres server in docker
+    """Start a PostgreSQL server in a Docker container.
+
+    If a PostgreSQL container is already running, returns the existing container.
+    Otherwise, starts a new container with the specified configuration.
+
     Args:
-        dbname (str): database name
-        dbuser (str): database user
-        dbpass (str): database password
-        version (str | None, optional): postgres version; defaults to "latest".
+        dbname (str, optional): Initial database name. Defaults to "timelink".
+        dbuser (str, optional): Database user. Defaults to "timelink".
+        dbpass (str, optional): Database password. If None, retrieves from environment.
+        image (str, optional): Docker image to use. Defaults to "postgres".
+        version (str, optional): Image version. Defaults to "latest".
+
+    Returns:
+        docker.models.containers.Container: The started or existing PostgreSQL container.
+
+    Raises:
+        RuntimeError: If Docker is not running or if the server fails to start/ready.
     """
     if not is_docker_running():
         raise RuntimeError("Docker is not running")
@@ -179,11 +246,23 @@ def start_postgres_server(
 
 
 def get_postgres_dbnames():
-    """Get the database names from a postgres server
-    Returns:
-        list[str]: list of database names
+    """Get the list of non-template databases from a running PostgreSQL server.
 
-    ..code-block:: sql
+    Queries the PostgreSQL system catalog to retrieve database names that:
+    - Are not templates (NOT datistemplate)
+    - Allow connections (datallowconn)
+    - Are not the 'postgres' system database
+
+    Returns:
+        list[str]: List of database names.
+
+    Raises:
+        RuntimeError: If Docker is not running.
+
+    Note:
+        This function will start a PostgreSQL server if one is not already running.
+
+    Example SQL query executed::
 
         SELECT datname
             FROM pg_database
@@ -213,7 +292,20 @@ def get_postgres_dbnames():
 
 
 def is_valid_postgres_db_name(db_name):
-    """Check if the database name is valid"""
+    """Validate a PostgreSQL database name according to PostgreSQL naming rules.
+
+    Args:
+        db_name (str): The database name to validate.
+
+    Returns:
+        bool: True if the name is valid, False otherwise.
+
+    Note:
+        A valid PostgreSQL database name must:
+        - Be less than 64 characters long
+        - Start with a letter or underscore
+        - Contain only letters, digits, and underscores
+    """
     # Check if the name is less than 64 characters long
     if len(db_name) >= 64:
         return False
